a lot of updates to how the navbar works
