+++ src/main/java/app/appl/customer/PostAddCreditCardRoute.java
@@ -1,10 +1,10 @@
-package app.util;
+package app.appl.customer;
 
 import spark.Request;
 import spark.Response;
 import spark.Route;
 
-public class GetBankAccountDataRoute implements Route {
+public class PostAddCreditCardRoute implements Route {
     @Override
     public Object handle(Request request, Response response) throws Exception {
         return null;
