+++ src/main/java/app/ui/customer/GetAddBankAccountRoute.java
@@ -1,12 +1,12 @@
-package app.ui;
+package app.ui.customer;
 
 import spark.Request;
 import spark.Response;
 import spark.Route;
 import spark.TemplateEngine;
 
-public class GetAddAddressRoute implements Route {
-    public GetAddAddressRoute(TemplateEngine templateEngine) {
+public class GetAddBankAccountRoute implements Route {
+    public GetAddBankAccountRoute(TemplateEngine templateEngine) {
     }
 
     @Override
