+++ src/main/java/app/util/customer/GetCreditCardDataRoute.java
@@ -1,10 +1,10 @@
-package app.appl;
+package app.util.customer;
 
 import spark.Request;
 import spark.Response;
 import spark.Route;
 
-public class PostAddCreditCardRoute implements Route {
+public class GetCreditCardDataRoute implements Route {
     @Override
     public Object handle(Request request, Response response) throws Exception {
         return null;
