+++ src/main/resources/spark/template/freemarker/admin/components/navbarContents.ftl
@@ -1,14 +1,6 @@
-<ul class="navbar-nav mr-auto">
-    <li class="nav-item active">
-        <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/tracking">Customers</a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/tracking">Delivery People</a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/account">Account</a>
-    </li>
-</ul>
+<li class="nav-item">
+    <a class="nav-link" href="/tracking">Customers</a>
+</li>
+<li class="nav-item">
+    <a class="nav-link" href="/tracking">Delivery People</a>
+</li>
