+++ src/main/resources/spark/template/freemarker/customer/components/navbarContents.ftl
@@ -1,11 +1,3 @@
-<ul class="navbar-nav mr-auto">
-    <li class="nav-item active">
-        <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/tracking">Tracking</a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/account">Account</a>
-    </li>
-</ul>
+<li class="nav-item">
+    <a class="nav-link" href="/tracking">Tracking</a>
+</li>
