+++ src/main/resources/spark/template/freemarker/delivery/components/navbarContents.ftl
@@ -1,14 +1,6 @@
-<ul class="navbar-nav mr-auto">
-    <li class="nav-item active">
-        <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/customers">Packages</a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/delivery">Delivery</a>
-    </li>
-    <li class="nav-item">
-        <a class="nav-link" href="/account">Account</a>
-    </li>
-</ul>
+<li class="nav-item">
+    <a class="nav-link" href="/customers">Packages</a>
+</li>
+<li class="nav-item">
+    <a class="nav-link" href="/delivery">Delivery</a>
+</li>
